# Switch to using gzip-hpp for compression in vtvalidate