--- a/.gitignore
+++ b/.gitignore
@@ -5,3 +5,4 @@ mason_packages
 .mason
 .toolchain
 local.env
+npm-debug.log