--- a/scripts/install_deps.sh
+++ b/scripts/install_deps.sh
@@ -14,3 +14,4 @@ source local.env
 
 install protozero 1.6.1
 install vtzero f6efb8e
+install gzip-hpp 0.1.0