## Description
Use gzip-hpp to decompress tile buffers for validation.

Per mapbox/core-tech#342 and https://github.com/mapbox/core-tech/issues/245
Referencing https://github.com/mapbox/vt-shaver-cpp/pull/152

## Changes
#### bin/vtvalidate.js 
* Removed `decompressBuffer` and now doing decompression in `src/vtvalidate.cpp`
#### scripts/install_deps.sh
* Updated vtzero dependency to 1.0.0
* Added gzip-hpp 0.1.0 dependency
#### scripts/setup.sh
* Updated mason to 1785c00
* Updated llvm to 6.0.0
#### src/vtvalidate.cpp
* Added decompression code using `gzip::Decompressor` to `AsyncValidateWorker`