@allieoop 2018-06-01
One test is currently failing. It expects a zlib compressed tile to be valid but we're getting back `invalid tag exception`.

This is the failing test:
```node
test('success: valid zlib compressed tile', function(t) {
  var tile = path.resolve(__dirname, 'fixtures', 'zlib-compressed');
  var cmd = [ validate, tile ].join(' ');

  exec(cmd, function(err, stdout, stderr) {
    t.ifError(err, 'no error');
    t.equal(stdout, '');
    t.end();
  });
});
```

This is the code that changed and is causing it to fail:
```c++
if (gzip::is_compressed(data.data(), data.size())) {
    gzip::Decompressor decompressor;
    std::string uncompressed;
    decompressor.decompress(uncompressed, data.data(), data.size());
    data = vtzero::data_view(uncompressed);
}
```

Which is very close to how we do decompression [here](https://github.com/mapbox/vt-shaver-cpp/blob/7b33efb1c062af2e1b137fc7e36e6bc7cb7d1fd4/src/shave.cpp#L330), expect we use a baton:
```c++
if (gzip::is_compressed(baton->data, baton->dataLength)) {
    // Decompress tile before reading data
    gzip::Decompressor decompressor;
    decompressor.decompress(uncompressed, baton->data, baton->dataLength);
    dv = vtzero::data_view(uncompressed);
}
```

So my next step is to try to understand why my uncompressed string looks like: `\x1a�\x01x\x02\n\rpark_features\x12\r\b\x01\x12\x02invalid tag exception` and what an `invalid tag exception` means and why/if using a baton is necessary.