@springmeyer 2018-06-01
Nice work pulling this together so quickly @allieoop.

As far as the error, I think the problem is likely:

```c++
data = vtzero::data_view(uncompressed);
```

As I suspect overwriting that data may be modifying it in a way that creates non-determinative results. How about:

```C++
            if (gzip::is_compressed(data.data(), data.size())) {
                gzip::Decompressor decompressor;
                std::string uncompressed;
                decompressor.decompress(uncompressed, data.data(), data.size());
                vtzero::data_view dv(uncompressed);
                result_ = parseTile(dv);
            } else {
                result_ = parseTile(data);
            }
```