@GretaCB 2018-06-04
@allieoop Awesome PR 🙌 . I just added a couple compression tests in isValid, since compression is now happening inside `vtvalidate.cpp`.