@allieoop 2018-06-05
Node 6 and 4 versions were published for linux and macOS: https://travis-ci.org/mapbox/vtvalidate/builds/387950731?utm_source=github_status&utm_medium=notification
via `./scripts/publish.sh` which did a `./node_modules/.bin/node-pre-gyp package publish $@`

Note: `publish` is used to publish binaries to s3 `mapbox-node-binary` bucket via commit messages if the commit message includes `[publish binary]`