@allieoop 2018-06-16
Even though these local benchmark tests seem to show a decrease in performance, you can see an improvement when mapbox-tile-copy uses this version of vtvalidate (benchmarks for mapbox-tile-copy are also below).

Using some real-world fixtures, we have the follow results:

## Before

* vtvalidate [uncompressed tiles](https://github.com/mapbox/mvt-fixtures/tree/master/real-world/chicago)

```
Benchmark speed: 9615 runs/s (runs: 1500 ms: 156 )
Benchmark peak mem (max_rss, max_heap, max_heap_total):  22.81MB 4.5MB 10.03MB
Benchmark iterations:  50 concurrency:  10
0.63s user 0.04s system 252% cpu 0.262 total
```

* vtvalidate [compressed tiles](https://github.com/mapbox/mvt-fixtures/tree/master/real-world/compressed)

```
Benchmark speed: 11111 runs/s (runs: 200 ms: 18 )
Benchmark peak mem (max_rss, max_heap, max_heap_total):  21.42MB 3.78MB 8.03MB
Benchmark iterations:  50 concurrency:  10
0.13s user 0.02s system 125% cpu 0.119 total
```

* vtvalidate in [mapbox-tile-copy](https://github.com/mapbox/mapbox-tile-copy)

```
52.51s user 38.78s system 146% cpu 1:02.41 total
```

## After 

* vtvalidate [uncompressed tiles](https://github.com/mapbox/mvt-fixtures/tree/master/real-world/chicago)

```
Benchmark speed: 3759 runs/s (runs: 1500 ms: 399 )
Benchmark peak mem (max_rss, max_heap, max_heap_total):  23.05MB 4.52MB 10.03MB
Benchmark iterations:  50 concurrency:  10
1.55s user 0.07s system 242% cpu 0.670 total
```

* vtvalidate [compressed tiles](https://github.com/mapbox/mvt-fixtures/tree/master/real-world/compressed)

```
Benchmark speed: 4878 runs/s (runs: 200 ms: 41 )
Benchmark peak mem (max_rss, max_heap, max_heap_total):  21.34MB 3.78MB 8.03MB
Benchmark iterations:  50 concurrency:  10
0.31s user 0.04s system 129% cpu 0.274 total
```

* vtvalidate in [mapbox-tile-copy](https://github.com/mapbox/mapbox-tile-copy)

```
32.70s user 32.45s system 160% cpu 40.678 total
```